Update jquery.typewatch.js
Very simple enhancement to allow one to do an alternative like remove an ok tick if the field is below the capture length. One does not want to want to show a certain field as eg. "correct" when the field was not even checked. So this will allow you to call "elsedo" and cleanup say a graphical ok tick to show field is ok.